[Finder] Disable failing test about open_basedir
